requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-dotenv==0.19.1
isort==5.10.1
black==21.10b0